Simplify st_write - remove update argumen, explain behaviour of overwrite

+++ R/read.R
@@ -134,8 +134,8 @@ clean_columns = function(obj, factorsAsCharacter) {
 #' @param quiet logical; suppress info on name, driver, size and spatial reference
 #' @param factorsAsCharacter logical; convert \code{factor} objects into character strings (default), else into numbers by
 #' \code{as.numeric}.
-#' @param update logical; if \code{TRUE}, try to update (append to) existing data source; this is only supported by some drivers
-#' @param overwrite logical; \code{FALSE} by default, if \code{TRUE} the original file will be overwritten by new data 
+#' @param overwrite logical; \code{FALSE} by default for single-layer data sources and \code{TRUE} by default for database data sources.
+#' If \code{TRUE} the original file will be overwritten by new data 
 #' (for drivers that write a single layer to a file) or an attempt will be made to update (append to) the existing data source
 #' (for database-type drivers defined by \code{db_drivers} such as GPKG).
 #' @details columns (variables) of a class not supported are dropped with a warning.
@@ -156,8 +156,10 @@ clean_columns = function(obj, factorsAsCharacter) {
 #' @export
 st_write = function(obj, dsn, layer = basename(dsn), driver = guess_driver_can_write(dsn), ...,
 		dataset_options = NULL, layer_options = NULL, quiet = FALSE, factorsAsCharacter = TRUE,
-		update = driver %in% db_drivers, overwrite = FALSE) {
+		overwrite = driver %in% db_drivers) {
 
+  is_db = driver %in% db_drivers
+  
 	if (length(list(...)))
 		stop(paste("unrecognized argument(s)", unlist(list(...)), "\n"))
 
@@ -168,18 +170,10 @@ st_write = function(obj, dsn, layer = basename(dsn), driver = guess_driver_can_w
 	if (missing(dsn))
 		stop("dsn should specify a data source or filename")
 	
-	if (file.exists(dsn) && overwrite){
-	        if (!(driver %in% db_drivers)){
-	                file.remove(dsn)   
-	        } else{
-	                stop("`overwrite` cannot be used for drivers that write a multi layer to a file")
-	        }
-	}
-	        
-	if (file.exists(dsn)) {
-	  dsn = normalizePath(dsn)
-	} else {
-	  dsn = path.expand(dsn)
+  dsn = normalizePath(dsn)
+	
+	if (file.exists(dsn) && overwrite && !is_db){
+	  file.remove(dsn)   
 	}
 		
 	geom = st_geometry(obj)
@@ -194,7 +188,7 @@ st_write = function(obj, dsn, layer = basename(dsn), driver = guess_driver_can_w
 			class(geom[[1]])[1]
 	CPL_write_ogr(obj, dsn, layer, driver,
 		as.character(dataset_options), as.character(layer_options),
-		geom, dim, quiet, update)
+		geom, dim, quiet, overwrite)
 }
 
 #' @name st_write
